Do not configure the fake policy evaluator by default
This way the application factory is more reusable as it does not make any assumptions about the application.
