+++ Services/CO.CDP.Forms.WebApi.Tests/Api/DeleteFormAnswerSetEndpointTest.cs
@@ -20,8 +20,10 @@ public class DeleteFormAnswerSetEndpointTest
         TestWebApplicationFactory<Program> factory = new(builder =>
         {
             builder.ConfigureServices(services =>
-                services.AddScoped(_ => _deleteAnswerSetUseCase.Object)
-            );
+            {
+                services.AddScoped(_ => _deleteAnswerSetUseCase.Object);
+                services.ConfigureFakePolicyEvaluator();
+            });
         });
         _httpClient = factory.CreateClient();
     }
