+++ Services/CO.CDP.Organisation.WebApi.Tests/Api/OrganisationEndpointsTests.cs
@@ -31,6 +31,7 @@ public class OrganisationEndpointsTests
                 services.AddScoped(_ => _getOrganisationsUseCase.Object);
                 services.AddScoped(_ => _getOrganisationUseCase.Object);
                 services.AddScoped(_ => _updatesOrganisationUseCase.Object);
+                services.ConfigureFakePolicyEvaluator();
             });
         });
         _httpClient = factory.CreateClient();
