+++ TestKit/CO.CDP.TestKit.Mvc/FakePolicyEvaluator.cs
@@ -1,0 +1,26 @@
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Authorization.Policy;
+using Microsoft.AspNetCore.Http;
+
+namespace CO.CDP.TestKit.Mvc;
+
+public class FakePolicyEvaluator : IPolicyEvaluator
+{
+    public async Task<AuthenticateResult> AuthenticateAsync(AuthorizationPolicy policy, HttpContext context)
+    {
+        var principal = new ClaimsPrincipal();
+
+        principal.AddIdentity(new ClaimsIdentity([new Claim("sub", "fake_sub")], "FakeScheme"));
+
+        return await Task.FromResult(AuthenticateResult.Success(new AuthenticationTicket(principal,
+            new AuthenticationProperties(), "FakeScheme")));
+    }
+
+    public async Task<PolicyAuthorizationResult> AuthorizeAsync(
+        AuthorizationPolicy policy, AuthenticateResult authenticationResult, HttpContext context, object? resource)
+    {
+        return await Task.FromResult(PolicyAuthorizationResult.Success());
+    }
+}
